Adding searchStudentById() updateStudentGradeById()
searchStudentById() is used to see if there is a student object in the
Student array with the same id value that is passed to
searchStudentById().

updateStudentGradeById() is used to update a grade of a student object.
It will first call searchStudentById() so to determine if the student id
wanting to be changed is in the Student array. If it is, then it will
find the student object and change the grade to be the double grade
value passed to the updateStudentGradeById() method.